refactor: move `ext_mod_impl::self` into `ext_mod_impl::lib`
